backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.13
langchain-chroma==0.1.0
sentence-transformers==2.2.2
chromadb==0.4.22
huggingface-hub==0.20.2
transformers==4.36.2
torch==2.1.2
beautifulsoup4==4.12.2
lxml==4.9.3